Added gcp connect, gcp update and fixed gcp list
