+++ src/commands/servers/list/gcp.ts
@@ -4,6 +4,7 @@ import chalk from "chalk";
 import { IInstancesData } from "../../../utils/interfaces";
 import { FlagInput } from "@oclif/core/lib/interfaces";
 import getGCP from "../../../utils/get/gcp";
+import { allGcpRegions } from "../../../utils/gcp-zones";
 
 export default class ListCommandGCP extends Command {
   static description = `GCP 
@@ -13,39 +14,10 @@ export default class ListCommandGCP extends Command {
   static flags: FlagInput<any> = {
     region: Flags.string({
       char: "r",
-      description: "Only get servers in a specific region(s)",
+      description: "Only get servers in a specific regions(s)",
       multiple: true,
       default: "all",
-      options: [
-        "asia-east1",
-        "asia-east2",
-        "asia-northeast1",
-        "asia-northeast2",
-        "asia-northeast3",
-        "asia-south1",
-        "asia-south2",
-        "asia-southeast1",
-        "asia-southeast2",
-        "australia-southeast1",
-        "europe-central2",
-        "europe-north1",
-        "europe-west1",
-        "europe-west2",
-        "europe-west3",
-        "europe-west4",
-        "europe-west6",
-        "northamerica-northeast1",
-        "northamerica-northeast2",
-        "southamerica-east1",
-        "southamerica-west1",
-        "us-central1",
-        "us-east1",
-        "us-east4",
-        "us-west1",
-        "us-west2",
-        "us-west3",
-        "us-west4"
-      ]
+      options: allGcpRegions
     }),
     state: Flags.string({
       char: "s",
@@ -67,8 +39,8 @@ export default class ListCommandGCP extends Command {
       multiple: true,
       default: "all"
     }),
-    "no-refresh": Flags.boolean({
-      description: "Don't refresh the cache of known servers",
+    "use-cache": Flags.boolean({
+      description: "Use the local instances cache file",
       default: false
     })
   };
@@ -88,8 +60,6 @@ export default class ListCommandGCP extends Command {
         chalk.blueBright("Index"),
         chalk.blueBright("Name"),
         chalk.blueBright("Address"),
-        chalk.blueBright("Key Pair"),
-        chalk.blueBright("Username"),
         chalk.blueBright("State"),
         chalk.blueBright("Accessible"),
         chalk.blueBright("Location"),
@@ -125,8 +95,6 @@ export default class ListCommandGCP extends Command {
           chalk.grey(`${instance.address}`) :
           chalk.white(`${instance.address}`)
         ),
-        "",
-        "",
         stateChalk,
         (instance.accessible === true ?
           chalk.green(`${instance.accessible}`) :
@@ -136,7 +104,7 @@ export default class ListCommandGCP extends Command {
           )
         ),
         instance.location,
-        "",
+        instance.account,
         "GCP"
       ]);
     }
