+++ src/utils/gcp-zones.ts
@@ -1,4 +1,4 @@
-export default function getZones(regions: string[]): string[] {
+export function getZones(regions: string[]): string[] {
   const regionList: any = {
     "asia-east1": [
       "asia-east1-a",
@@ -61,7 +61,6 @@ export default function getZones(regions: string[]): string[] {
       "europe-north1-c"
     ],
     "europe-west1": [
-      "europe-west1-a",
       "europe-west1-b",
       "europe-west1-c"
     ],
@@ -149,3 +148,34 @@ export default function getZones(regions: string[]): string[] {
 
   return zones;
 }
+
+export const allGcpRegions = [
+  "asia-east1",
+  "asia-east2",
+  "asia-northeast1",
+  "asia-northeast2",
+  "asia-northeast3",
+  "asia-south1",
+  "asia-south2",
+  "asia-southeast1",
+  "asia-southeast2",
+  "australia-southeast1",
+  "europe-central2",
+  "europe-north1",
+  "europe-west1",
+  "europe-west2",
+  "europe-west3",
+  "europe-west4",
+  "europe-west6",
+  "northamerica-northeast1",
+  "northamerica-northeast2",
+  "southamerica-east1",
+  "southamerica-west1",
+  "us-central1",
+  "us-east1",
+  "us-east4",
+  "us-west1",
+  "us-west2",
+  "us-west3",
+  "us-west4"
+];
