+++ src/utils/get/gcp.ts
@@ -2,7 +2,7 @@ import Compute, { InstancesClient } from "@google-cloud/compute";
 import { Config } from "@oclif/core";
 import { FlagInput } from "@oclif/core/lib/interfaces";
 import chalk from "chalk";
-import getZones from "../get-zones";
+import { allGcpRegions, getZones } from "../gcp-zones";
 import { IConfigData, IInstance, IInstancesData } from "../interfaces";
 import { isPortReachable, readJsonFile, writeJsonFile } from "../utils";
 
@@ -19,7 +19,7 @@ export default async function getGCP(flags: FlagInput<any>, config: Config): Pro
     return;
   }
 
-  if (JSON.parse(flags["no-refresh"].toString())) {
+  if (JSON.parse(flags["use-cache"].toString())) {
     const instances: IInstancesData = {
       aws: [],
       gcp: instancesData.gcp.filter((instance: IInstance) => {
@@ -37,9 +37,10 @@ export default async function getGCP(flags: FlagInput<any>, config: Config): Pro
   console.log(`${chalk.green("[INFO]")} Gathering GCP managed instances`);
 
   for await (const account of configData.gcpAccounts) {
+    console.log(`${chalk.green("[INFO]")} Checking account: ${account.gcpAccountName}`);
     process.env.GOOGLE_APPLICATION_CREDENTIALS = account.credentialsFile;
 
-    flags.region.toString() === "all" ? instancesData.gcp = await listAllInstances(account.gcpProjectId) : instancesData.gcp = await listInstanceByZone(account.gcpProjectId, flags.region.toString().split(","));
+    flags.region.toString() === "all" ? instancesData.gcp = await listInstanceByZone(account.gcpProjectId, allGcpRegions) : instancesData.gcp = await listInstanceByZone(account.gcpProjectId, flags.region.toString().split(","));
   }
 
   const writeResponse: boolean = await writeJsonFile(config.dataDir, "instances", JSON.stringify(instancesData));
@@ -51,53 +52,21 @@ export default async function getGCP(flags: FlagInput<any>, config: Config): Pro
   return instancesData;
 }
 
-async function listAllInstances(projectId: string): Promise<any> {
-  const instanceClient: InstancesClient = new Compute.InstancesClient();
-  const instancesData: any = [];
-  let ephermeralIP: string = "";
-  let accessible: boolean = false;
-  const instanceList = instanceClient.aggregatedListAsync({
-    project: projectId,
-    maxResults: 5
-  });
-
-  for await (const [zone, instancesObject] of instanceList) {
-    const instances = instancesObject.instances;
-
-    if (instances && instances.length > 0) {
-      for await (const instance of instances) {
-        if (!instance.networkInterfaces || instance.networkInterfaces.length < 0) {
-          return;
-        }
-
-        if (!instance.networkInterfaces[0].accessConfigs || instance.networkInterfaces[0].accessConfigs.length < 0) {
-          ephermeralIP = "Unknown";
-        } else if (instance.networkInterfaces[0].accessConfigs.length > 0 && instance.networkInterfaces[0].accessConfigs[0].natIP) {
-          ephermeralIP = instance.networkInterfaces[0].accessConfigs[0].natIP;
-        } else {
-          ephermeralIP = "Unknown";
-        }
-
-        if (ephermeralIP !== "Unknown") {
-          accessible = await isPortReachable(22, { host: ephermeralIP, timeout: 1000 });
-        }
-
-        instancesData.push({ name: instance.name, address: ephermeralIP, username: "", hasKeyPair: "", keyPair: "", state: instance.status, accessible: accessible, location: zone, account: "" });
-      }
-    }
-  }
-
-  return instancesData;
-}
-
 async function listInstanceByZone(projectId: string, regions: string[]): Promise<any> {
   const instanceClient: InstancesClient = new Compute.InstancesClient();
   const instancesData: any = [];
   let ephermeralIP: string = "";
   let accessible: boolean = false;
-  const zones = getZones(regions);
+  const zonesToSearch = getZones(regions);
+  const searchedRegions: string[] = [];
 
-  for await (const zone of zones) {
+  for await (const zone of zonesToSearch) {
+    const inRegion = zone.replace("zones/", "").replace(/-\w$/, "");
+    if (!searchedRegions.includes(inRegion)) {
+      console.log(`${chalk.green("[INFO]")} Checking region: ${inRegion}`);
+      searchedRegions.push(inRegion);
+    }
+
     const instanceList = instanceClient.listAsync({
       project: projectId,
       zone: zone
@@ -120,7 +89,14 @@ async function listInstanceByZone(projectId: string, regions: string[]): Promise
         accessible = await isPortReachable(22, { host: ephermeralIP, timeout: 1000 });
       }
 
-      instancesData.push({ name: instance.name, address: ephermeralIP, username: "", hasKeyPair: "", keyPair: "", state: instance.status, accessible: accessible, location: zone, account: "" });
+      instancesData.push({
+        name: instance.name,
+        address: ephermeralIP,
+        state: instance.status,
+        accessible: accessible,
+        location: zone.replace("zones/", "").replace(/-\w$/, ""),
+        account: projectId
+      });
     }
   }
